+++ backend/tests/test_integration_engagement.py
@@ -2049,3 +2049,26 @@ class Tests(util.LoRATestCase):
                 },
             },
         )
+
+    def test_reading_engagement_only_primary_uuid(self):
+        self.load_sample_structures()
+
+        self.assertRequestResponse(
+            '/service/e/53181ed2-f1de-4c4a-a8fd-ab358c2c454a'
+            '/details/engagement?only_primary_uuid=1',
+            [{
+                'engagement_type': {
+                    'uuid': '32547559-cfc1-4d97-94c6-70b192eff825'
+                },
+                'fraction': None,
+                'job_function': {
+                    'uuid': '4311e351-6a3c-4e7e-ae60-8a3b2938fbd6'
+                },
+                'org_unit': {'uuid': '9d07123e-47ac-4a9a-88c8-da82e3a4bc9e'},
+                'person': {'uuid': '53181ed2-f1de-4c4a-a8fd-ab358c2c454a'},
+                'primary': None,
+                'user_key': 'bvn',
+                'uuid': 'd000591f-8705-4324-897a-075e3623f37b',
+                'validity': {'from': '2017-01-01', 'to': None}
+            }]
+        )