+++ backend/mora/service/detail_writing.py
@@ -136,6 +136,8 @@ def create():
     :<jsonarr boolean primary: Mark this one as the “main” engagement
                                or position of a given employee. There
                                can only be one.
+    :<jsonarr integer fraction: An indication of how much this
+        engagement constitutes the employee's overall employment
     :<jsonarr object job_function: The job function of the association
     :<jsonarr object engagement_type: The engagement type
     :<jsonarr string user_key: Short, unique key identifying the relation.
@@ -497,6 +499,10 @@ def edit():
     :<jsonarr object job_function: The job function of the association
     :<jsonarr object engagement_type: The engagement type
     :<jsonarr object validity: The validities of the changes.
+    :<jsonarr boolean primary: Whether this is the one and only main
+                               position for the relevant person.
+    :<jsonarr integer fraction: An indication of how much this
+        engagement constitutes the employee's overall employment
 
     The parameters ``job_function`` and ``engagement_type`` should contain
     UUIDs obtained from their respective facet endpoints.